requirements.txt
# Dependencias de Python para el extractor de QR desde PDF
PyMuPDF
pyzbar
Pillow
opencv-python
numpy